requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
idna==3.2
Pillow==8.3.2
python-resize-image==1.1.19
qrcode==7.3
requests==2.26.0
urllib3==1.26.6
